ros_control: renamed dynamixel hardware interface
